Combine old classes to provide better starting point for migration
